+++ src/novita/api/gpu.py
@@ -12,7 +12,6 @@ from .resources.gpu import (
     AsyncNetworks,
     AsyncProducts,
     AsyncRegistries,
-    AsyncSSHKeys,
     AsyncStorages,
     AsyncTemplates,
     Clusters,
@@ -24,7 +23,6 @@ from .resources.gpu import (
     Networks,
     Products,
     Registries,
-    SSHKeys,
     Storages,
     Templates,
 )
@@ -52,7 +50,6 @@ class GpuClient:
         self.networks = Networks(client)
         self.products = Products(client)
         self.registries = Registries(client)
-        self.ssh_keys = SSHKeys(client)
         self.storages = Storages(client)
         self.templates = Templates(client)
 
@@ -76,6 +73,5 @@ class AsyncGpuClient:
         self.networks = AsyncNetworks(client)
         self.products = AsyncProducts(client)
         self.registries = AsyncRegistries(client)
-        self.ssh_keys = AsyncSSHKeys(client)
         self.storages = AsyncStorages(client)
         self.templates = AsyncTemplates(client)